Separate hip/cuda buffer into their own headers
Also contains a fix for the aggregation failure test
(just re-enabling the test by defining
DEBUG_AGGREGATION_CALLS within the test itself)
